Move TokenType from types.js to lexer.js and update imports
